added basic search frontend components
