+++ frontend/components/search/search_results.jsx
@@ -1,0 +1,12 @@
+import React, { Component } from 'react'
+
+export default class SearchResults extends Component {
+	render() {
+		
+		return (
+			<div>
+				Hello.
+			</div>
+		)
+	}
+}
